Resized Welcome shield, added insert USB graphic, minor copy change
I made the Welcome shield too big when I changed it to SVG, I’ve
altered that page a little bit so it all fits again.

Minor copy change on Welcome6 to reflect the new badge menu.

Included a graphic for the “insert usb” page, although I’m not sure
we’re using it any more.